requirements.txt
Django
django-anymail[sendgrid,sparkpost]
django-front
djangorestframework
django-environ